test(senpi-task): exercise async notification chaos
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
