+++ packages/senpi-task/src/__adversarial__/chaos-harness.ts
@@ -14,8 +14,8 @@ import { ChaosProcessTable, ChaosRunner } from "./chaos-engine"
 import type { LifecycleChaosObservations } from "./chaos-engine"
 import { buildChaosEngines } from "./chaos-engine-factory"
 import type { ChaosEngine } from "./chaos-engine-factory"
-import { createChaosNotifier, createNotificationEpochTracker, instrumentCompletionNotifier } from "./chaos-invariants"
-import type { ChaosNotifier } from "./chaos-invariants"
+import { createChaosNotifier, createNotificationEpochTracker, instrumentCompletionNotifier } from "./chaos-notifications"
+import type { ChaosNotifier } from "./chaos-notifications"
 import { createObservingStore } from "./observing-store"
 import type { StoreObservations } from "./observing-store"
